Update README.md
Clarified the description on how the arguments to Range() influence the generate the resulting argument sequence.

This should make things less confusing for new users of the library, since the exact mechanics are currently nowhere clearly specified and require reading the source code in `benchmark_register.cc` to figure out.